Actor creation is injectable.
Debug sponsor is 'control' sponsor. Added mechanism for injecting
how configuration should be constructed. README updates. Removed
previous tracing from core and put it in 'examples/tracing.js'.

The tests that were tied to tart.tracing are now using the
'examples/tracing.js' example. This should probably be refactored
to a separate module in the future.